+++ src/settings/descriptors/staleDataDescriptor.ts
@@ -28,7 +28,7 @@ import { BaseDescriptor } from "./baseDescriptor";
 
 export class StaleDataDescriptor extends BaseDescriptor {
     public staleDataText: string = undefined;
-    public staleDataCriteria: number = undefined;
+    public staleDataThreshold: number = 0;
     public color: string = "#3599b8";
     public background: string = "";
 }