+++ src/settings/settings.ts
@@ -98,10 +98,8 @@ export class Settings extends SettingsBase {
     }
 
     public parse(): void {
-        if (this.subtitle.staleDataText
-            && this.staleData.staleDataText === undefined
-        ) {
-            this.staleData.staleDataText = "Data is ${1} days old." + this.subtitle.staleDataText;
+        if (this.staleData.staleDataText === undefined) {
+            this.staleData.staleDataText = "Data is ${1} days old." + (this.subtitle.staleDataText || "");
         }
 
         if (!this.subtitle.shouldBeShown) {